requirements.txt
absl-py==1.4.0
asttokens 
astunparse==1.6.3
backcall
backports.functools-lru-cache 
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama 
debugpy 
decorator 
executing 
Flask==2.3.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata 
ipykernel 
ipython 
itsdangerous==2.1.2
jax==0.4.12
jedi 
Jinja2==3.1.2
jupyter_client 
jupyter_core 
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline 
ml-dtypes==0.2.0
nest-asyncio
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging 
parso 
pickleshare
pillow
platformdirs 
prompt-toolkit 
protobuf==4.23.2
psutil
pure-eval 
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments 
python-dateutil 
pyzmq 
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six 
stack-data 
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tornado 
traitlets 
typing_extensions 
urllib3==1.26.16
wcwidth
Werkzeug==2.3.6
wrapt==1.14.1
zipp 